% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iMIRAGE.R
\name{CVProc}
\alias{CVProc}
\title{Process cross-validation results}
\usage{
CVProc(res)
}
\arguments{
\item{res}{The output object from imirage.cv}
}
\value{
a processed matrix containing 3 columns: Spearman's correlation coefficient, P-value and root mean
squared error from cross-validation analysis
}
\description{
This function returns useful information by organizing the output from cross-validation analysis.
Used internatlly by imirage.cv.loop and can be used for processing output from imirage.cv
}
