% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.exp.R
\name{filter.exp}
\alias{filter.exp}
\title{Pre-standardization filter to select miRNA expressed in a specific proportion of samples}
\usage{
filter.exp(train_mir, cutoff = 75, threshold = 0)
}
\arguments{
\item{train_mir}{an expression matrix with miRNA in columns, samples in rows}

\item{cutoff}{percentage of samples in which the miRNA should be expressed}

\item{threshold}{the numeric threshold defining expression of miRNA. Default threshold = 0.}
}
\value{
filtered miRNA expression matrix
}
\description{
return a subset of the miRNA expression dataset containing miRNAs that are expressed
in at least a specified percentage of samples based on a user-defined expression threshold
}
\examples{
filter.exp(train_mir, cutoff=75, threshold=0)

}
