% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iMIRAGE.R
\name{imirage.cv.loop}
\alias{imirage.cv.loop}
\title{iMIRAGE cross-validation loop function for full miRNA matrix}
\usage{
imirage.cv.loop(train_pcg, train_mir, method = "KNN", ...)
}
\arguments{
\item{train_pcg}{training protein coding dataset. a numeric matrix with with row names indicating
samples, and column names indicating protein coding gene IDs.}

\item{train_mir}{training miRNA expression dataset. a numeric matrix with row names indicating
samples, and column names indicating miRNA IDs.}

\item{method}{method for imputation, either "RF" for random forests, "KNN" for K-nearest neighbor or
"SVM" for support vector machines.}

\item{...}{optional parameters that can be passed on to the machine-learning method:
RF (\link[randomForest]{randomForest}), KNN (\link[FNN]{knn.reg}) or SVM(\link[e1071]{svm})}
}
\value{
a matrix containing Spearman's correlation coefficient, P-value and RMSE from the cross-validation analysis
of the complete miRNA training dataset
}
\description{
Convinient wrapper for \link[iMIRAGE]{imirage.cv} that performs cross-validation analysis for
assessing imputation accuracies for all miRNAs using the training datasets
}
\examples{

}
