% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imirage.cv.R
\name{imirage.cv}
\alias{imirage.cv}
\title{Cross validation function for iMIRAGE imputation accuracy}
\usage{
imirage.cv(train_pcg, train_mir, gene_index, num = 50, method = "KNN",
  folds = 10, target = "none", ...)
}
\arguments{
\item{train_pcg}{training protein coding dataset. a numeric matrix with row names indicating
samples, anSed column names indicating protein coding gene IDs.}

\item{train_mir}{training miRNA expression dataset. a numeric matrix with row names indicating
samples, and column names indicating miRNA IDs}

\item{gene_index}{either gene name (character) or index (column number) of miRNA to be imputed.}

\item{num}{number of informative protein coding genes to be used in constructing imputation model.
Default is 50 genes.}

\item{method}{method for imputation, either "RF" for random forests, "KNN" for K-nearest neighbor or
"SVM" for support vector machines.}

\item{folds}{number specifying folds (k) of cross validation to obtain imputation accuracy.
Default is k=10.}

\item{target}{"none" (default), "ts.pairs", or dataframe/matrix/list.
this argument accepts character strings to indicate the use of all candidate genes as predictors ("none),
or use built-in TargetScan miRNA-gene pairs ("ts.pairs"). also accepts a dataframe , matrix or list object
containing a column with names of miRNA and a column with the names of target genes.}

\item{...}{optional parameters that can be passed on to the machine-learning functions
RF (\link[randomForest]{randomForest}), KNN (\link[FNN]{knn.reg}) or SVM(\link[e1071]{svm})}
}
\value{
a matrix with three values corresponding to Spearman's correlation coefficient,
P-value of the fit and root mean squared error (RMSE).
}
\description{
Function to obtain accuracy parameters: correlation coefficient, P-value and RMSE of
imputation model
}
\examples{
data(iMIRAGE.datasets)
imirage.cv(GA.pcg, GA.mir, gene_index="hsa-let-7c", method="KNN", num=50)
imirage.cv(GA.pcg, GA.mir, gene_index=25, method="KNN", num=50)

}
