% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iMIRAGE.R
\name{imirage}
\alias{imirage}
\title{iMIRAGE: imputated miRNA Activity from Gene Expression}
\usage{
imirage(train_pcg, train_mir, my_pcg, gene_index, method = "KNN",
  num = 50, target = FALSE, ...)
}
\arguments{
\item{train_pcg}{training protein coding dataset. a numeric matrix with with row names indicating
samples, and column names indicating protein coding gene IDs.}

\item{train_mir}{training miRNA expression dataset. a numeric matrix with row names indicating
samples, and column names indicating miRNA IDs.}

\item{my_pcg}{test protein coding expression dataset. a numeric matrix with row names indicating
samples, and column names indicating protein coding gene IDs.}

\item{gene_index}{either gene name (character) or index (column number) of miRNA to be imputed.}

\item{method}{method for imputation, either "RF" for random forests, "KNN" for K-nearest neighbor or
"SVM" for support vector machines. Uses KNN by default.}

\item{num}{number of informative protein coding genes to be used in constructing imputation model.
Default is 50 genes.}

\item{target}{logical, specifying whether protein coding genes should be restricted to predicted
targets of the miRNA (from TargetScan) or use all genes as candidates. Default = FALSE.}

\item{...}{optional parameters that can be passed on to the machine-learning method:
RF (\link[randomForest]{randomForest}), KNN (\link[FNN]{knn.reg}) or SVM(\link[e1071]{svm})}
}
\value{
a numeric vector with imputed expression levels of the miRNA
}
\description{
Function to impute miRNA expression profile from protein coding expression dataset
}
\examples{
\\code{
data(iMIRAGE.datasets)
imirage(GA.pcg, GA.mir, HS.pcg, gene_index="hsa-let-7c", method="KNN", num=50)
imirage(GA.pcg, GA.mir, HS.pcg, gene_index=25, method="KNN", num=50)
}

}
