% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iMIRAGE.R
\name{match_mat}
\alias{match_mat}
\title{Match matrices by gene ID}
\usage{
match_mat(train_pcg, my_pcg)
}
\arguments{
\item{train_pcg}{a numeric matrix. training protein coding dataset to construct miRNA imputation model. Gene IDs are specified in column names.}

\item{my_pcg}{a numeric matrix. test protein coding dataset to impute miRNA profile. Gene IDs are specified in column names.}
}
\value{
list containing two matrices with common protein coding genes
}
\description{
Function to return matched matrices with common genes
}
\examples{
temp <- match_mat(train_pcg, my_pcg)
new_train_pcg <- data.matrix(temp[[1]])
new_my_pcg <- data.matrix(temp[[2]])

}
