% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iMIRAGE.R
\name{mirage}
\alias{mirage}
\title{MIRAGE: miRNA Activity from Gene Expression}
\usage{
mirage(train_pcg, train_mir, my_pcg, gene_index, num = 100)
}
\arguments{
\item{train_pcg}{training protein coding dataset. a numeric matrix with with row names indicating samples, and column names indicating protein coding gene IDs.}

\item{train_mir}{training miRNA expression dataset. a numeric matrix with row names indicating samples, and column names indicating miRNA IDs}

\item{my_pcg}{test protein coding expression dataset. a numberic matrix with row names indicating samples, and column names indicating protein coding gene IDs.}

\item{gene_index}{either gene name (character) or index (column number) of miRNA to be imputed.}

\item{num}{number of informative protein coding genes to be used in constructing imputation model. Default is 100 genes.}
}
\value{
a numeric vector containing imputed & standardized expression levels of the miRNA
}
\description{
Function to impute miRNA expression profile from protein coding expression dataset
}
\examples{
mirage(train_pcg, train_mir, my_pcg, gene_index="ENSG00000228630", num=100)
mirage(train_pcg, train_mir, my_pcg, gene_index=25, num=100)

}
