% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iMIRAGE.R
\name{mirage_cv}
\alias{mirage_cv}
\title{Cross validation function for imputation accuracy}
\usage{
mirage_cv(train_pcg, train_mir, gene_index, num = 100, folds = 5)
}
\arguments{
\item{train_pcg}{training protein coding dataset. a numeric matrix with row names indicating samples, and column names indicating protein coding gene IDs.}

\item{train_mir}{training miRNA expression dataset. a numeric matrix with row names indicating samples, and column names indicating miRNA IDs}

\item{gene_index}{either gene name (character) or index (column number) of miRNA to be imputed.}

\item{num}{number of informative protein coding genes to be used in constructing imputation model. Default is 100 genes.}

\item{folds}{number specifying folds of cross validation to obtain imputation accuracu. Default is 5.}
}
\value{
a matrix with three values corresponding to Pearson's correlation coefficient, P-value of fit and root mean square error
}
\description{
Function to obtain accuracy parameters: correlation coefficient, P-value and RMSE of imputation model
}
\examples{
mirage_cv(train_pcg, train_mir, gene_index="ENSG00000184441", num=100)
mirage_cv(train_pcg, train_mir, gene_index=25, num=100)

}
