% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iMIRAGE.R
\name{pre.process}
\alias{pre.process}
\title{Pre-processing gene expression matrix}
\usage{
pre.process(gex, var.filter = TRUE, log = FALSE, UQ = FALSE, std = TRUE)
}
\arguments{
\item{gex}{a gene expression matrix with genes in columns and samples in rows}

\item{var.filter}{logical. specify whether genes are filtered based on variance. Default = TRUE}

\item{log}{logical. specify whether log2(x+1) transformation should be performed. Default = FALSE}

\item{UQ}{logical. specify whether upper quantile normalization should be performed. Default = FALSE}

\item{scale}{logical. specifiy whether standardization should be performed. Default = TRUE}
}
\value{
a processed gene expression matrix
}
\description{
Use this function to perform a number of pre-processing steps, including 
log transformation, normalization and standardization of gene expression matrix. 
NOTE: If multiple options are set true, the sequence of pre-processing steps
performed are variance filtering, log2 transformation, upper quantile normalization 
and standardization.
}
\examples{
train_pcg <- pre.process(train_pcg)

}
