% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iMIRAGE.R
\name{posexp}
\alias{posexp}
\title{Filter miRNA dataset based on positive expression in a specific percentage of samples}
\usage{
posexp(train_mir, cutoff = 75, threshold = 1)
}
\arguments{
\item{train_mir}{a numeric matrix with miRNA expression data}

\item{cutoff}{percentage of samples in which the miRNA should be expressed}

\item{threshold}{the numeric threshold that defines whether the miRNA is expressed or not. Defaults to 1 (for RPKM or FPKM)}
}
\value{
filtered miRNA expression matrix
}
\description{
Function to return miRNA expression subset containing miRNAs that are expressed in a specific percentage of samples
}
\examples{
posexp(train_mir, cutoff=75, threshold=1)
}
